+++ CHANGELOG.md
@@ -9,7 +9,8 @@ ByteriX, 2017. All right reserved.
 ## 0.9.29 (07.02.2017)
 ##### Improvements
 * added iconMode to BxInputSelectorPicturesRow
-* animated adding pictures
+* animated adding of a pictures
+* improvement adding animation
 
 ## 0.9.28 (06.02.2017)
 ##### Bug fixing